+++ iosHyperskillApp/iosHyperskillApp/Sources/Modules/StepQuiz/StepQuizViewModel.swift
@@ -50,67 +50,3 @@ extension StepQuizViewModel: StepQuizChildQuizDelegate {
         self.doMainQuizAction()
     }
 }
-
-extension Reply {
-    convenience init(
-        sortingChoices: [Bool]? = nil,
-        tableChoices: [TableChoiceAnswer]? = nil,
-        text: String? = nil,
-        attachments: [Attachment]? = nil,
-        formula: String? = nil,
-        number: String? = nil,
-        ordering: [Int]? = nil,
-        language: String? = nil,
-        code: String? = nil,
-        blanks: [String]? = nil,
-        solveSql: String? = nil
-    ) {
-        let choicesAnswer: [ChoiceAnswer]? = {
-            if let sortingChoices = sortingChoices {
-                return sortingChoices.map(ChoiceAnswerChoice.init(boolValue:))
-            } else if let tableChoices = tableChoices {
-                return tableChoices.map(ChoiceAnswerTable.init(tableChoice:))
-            }
-            return nil
-        }()
-
-        self.init(
-            choices: choicesAnswer,
-            text: text,
-            attachments: attachments,
-            formula: formula,
-            number: number,
-            ordering: ordering?.map({ KotlinInt(value: Int32($0)) }),
-            language: language,
-            code: code,
-            blanks: blanks,
-            solveSql: solveSql
-        )
-    }
-}
-
-extension Dataset {
-    convenience init(
-        options: [String]? = nil,
-        pairs: [Pair]? = nil,
-        rows: [String]? = nil,
-        columns: [String]? = nil,
-        desc: String? = nil,
-        components: [Component]? = nil,
-        isMultipleChoice: Bool = false,
-        isCheckbox: Bool = false,
-        isHtmlEnabled: Bool = false
-    ) {
-        self.init(
-            options: options,
-            pairs: pairs,
-            rows: rows,
-            columns: columns,
-            description: desc,
-            components: components,
-            isMultipleChoice: isMultipleChoice,
-            isCheckbox: isCheckbox,
-            isHtmlEnabled: isHtmlEnabled
-        )
-    }
-}
